pyproject.toml
[tool.poetry]
name="testlitestar"
version="1.0"
description="LiteStar test task"
authors=[]

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1"
litestar = { extras = ["standard"], version = "^2" }
litestar-granian = "^0"
litestar-asyncpg = "^0"
advanced-alchemy = "^0.20"
msgspec = "^0.18.6"
alembic = "^1.15.2"
